Pull request 281: use hostsfile
Merge in GO/dnsproxy from use-hostsfile to master

Squashed commit of the following:

commit 4b3e59be00af2037c0e1f1213313dcc669a9baab
Author: Eugene Burkov <E.Burkov@AdGuard.COM>
Date:   Tue Sep 19 20:13:29 2023 +0300

    all: imp code

commit 9429f585b8fd39ad52e80f2c8c4dd4ed56266106
Author: Eugene Burkov <E.Burkov@AdGuard.COM>
Date:   Fri Sep 15 17:02:36 2023 +0300

    main: fix merge

commit f929b3c0f6c86c36eaaa3d9389165d509a432976
Merge: c442695 930f37e
Author: Eugene Burkov <E.Burkov@AdGuard.COM>
Date:   Fri Sep 15 17:00:29 2023 +0300

    Merge branch 'master' into use-hostsfile

commit c442695b2eb6a50435f1eef4aaec6e9f6847e598
Author: Eugene Burkov <E.Burkov@AdGuard.COM>
Date:   Fri Sep 15 16:41:16 2023 +0300

    netutil: fix lint

commit db0755113aad9b28118461c2594cd58a281c366e
Author: Eugene Burkov <E.Burkov@AdGuard.COM>
Date:   Fri Sep 15 16:37:33 2023 +0300

    all: imp code

commit cc76b8deaf05108d30fc0b585a8588fe09a2b239
Author: Eugene Burkov <E.Burkov@AdGuard.COM>
Date:   Thu Sep 14 20:57:42 2023 +0300

    netutil: fix windows again

commit 1d78807837a6f7bc780392f99955012bb72d3919
Author: Eugene Burkov <E.Burkov@AdGuard.COM>
Date:   Thu Sep 14 20:45:44 2023 +0300

    upstream: rm todo

commit b0125a3d7793e1fd50904c418dff4c0b14d30b9a
Author: Eugene Burkov <E.Burkov@AdGuard.COM>
Date:   Thu Sep 14 20:38:41 2023 +0300

    netutil: fix windows tests

commit 1f5e6c436d1b546bcda0e3dbb67f1a1c80f2abd1
Author: Eugene Burkov <E.Burkov@AdGuard.COM>
Date:   Thu Sep 14 15:06:12 2023 +0300

    all: imp code

commit d750be5c051591a902b89ca2acabca30a94e277d
Author: Eugene Burkov <E.Burkov@AdGuard.COM>
Date:   Thu Sep 14 12:19:37 2023 +0300

    all: imp code, docs

commit ca395ab88d9a171551327c088a9096833fc4eaa1
Author: Eugene Burkov <E.Burkov@AdGuard.COM>
Date:   Thu Sep 14 02:18:33 2023 +0300

    all: imp code

commit 7c1e5558a9e93fe751e6c1d78496e9ae30085969
Author: Eugene Burkov <E.Burkov@AdGuard.COM>
Date:   Tue Sep 12 14:26:41 2023 +0500

    all: imp code

commit 2cc8adf24b675eb7083be5cb4c492789e62e71d8
Author: Eugene Burkov <E.Burkov@AdGuard.COM>
Date:   Thu Sep 7 16:12:01 2023 +0300

    netutil: fix import

commit 519c1d09d96033096397d8fe0d97a4720d034456
Author: Eugene Burkov <E.Burkov@AdGuard.COM>
Date:   Thu Sep 7 16:05:45 2023 +0300

    all: add hostsresolver

commit b1c17943dd37172384e40ae5fa1bbee366115b63
Author: Eugene Burkov <E.Burkov@AdGuard.COM>
Date:   Mon Aug 28 16:00:05 2023 +0300

    netutil: add hosts

+++ upstream/upstream_test.go
@@ -416,7 +416,7 @@ func TestAddPort(t *testing.T) {
 		name string
 		want string
 		host string
-		port int
+		port uint16
 	}{{
 		name: "empty",
 		want: ":0",
